Bugfix: satvalues - position must be float in all steps
git-svn-id: https://svn.met.no/diana/trunk@2994 0ccb8871-b436-41a3-ae7d-e3ac59973202

+++ src/diController.cc
@@ -397,7 +397,7 @@ vector<miString> Controller::getCalibChannels(){
 }
 
 // show values in grid position x,y
-vector<SatValues> Controller::showValues(int x, int y){
+vector<SatValues> Controller::showValues(float x, float y){
   return plotm->showValues(x,y);
 }
